// frontend/app/lib/api/generated/models/cpuMemoryOverview.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Internal Server Observation
 * OpenAPI spec version: 0.1.0
 */

export interface CpuMemoryOverview {
  cpu_count: number;
  load_average_15m: number;
  load_average_1m: number;
  load_average_5m: number;
  memory_available: string;
  memory_free: string;
  memory_total: string;
  memory_used: string;
  memory_used_percent: number;
  swap_free: string;
  swap_total: string;
  swap_used: string;
  swap_used_percent: number;
}